Replace old group authentication with SAML SSO
In particular, this will change our login endpoints to bounce users to
SAML SSO, in order to grab some basic information about the user.

This is the most basic integration, we only store the locally unique
identifier off the response hash, from the OmniAuth schema.

More information on the response hash we're passed can be found on the
OmniAuth wiki[1]

This requires that the configuration XML file be present locally, and
that the location of that file is set in the `SAML_XML_LOCATION` envvar.

[1]: https://github.com/intridea/omniauth/wiki/Auth-Hash-Schema